+++ clara/transpiler/naive_bayes/__init__.py
@@ -1,3 +1,5 @@
 from.gausianNB import GaussianNBTranspiler
 from .complementNB import ComplementNBTranspiler
 from .multiNB import MultinomialNBTranspiler
+from .categoricalNB import CatergoricalNBTranspiler
+from .bernoulliNB import BernoulliNBTranspiler